Split key_value sort into a separate test
