+++ sycl/test-e2e/GroupAlgorithm/SYCL2020/group_sort/group_and_joint_sort.cpp
@@ -29,6 +29,7 @@
 
 #include <sycl/detail/core.hpp>
 
+#include "common.hpp"
 #include <sycl/builtins.hpp>
 #include <sycl/ext/oneapi/experimental/group_sort.hpp>
 #include <sycl/group_algorithm.hpp>
@@ -38,31 +39,6 @@
 #include <numeric>
 #include <random>
 #include <vector>
-
-namespace oneapi_exp = sycl::ext::oneapi::experimental;
-
-template <typename...> class KernelNameOverGroup;
-template <typename...> class KernelNameJoint;
-template <typename...> class KernelNameKeyValueOverGroup;
-
-enum class UseGroupT { SubGroup = true, WorkGroup = false };
-
-// these classes are needed to pass non-type template parameters to KernelName
-template <int> class IntWrapper;
-template <UseGroupT> class UseGroupWrapper;
-
-class CustomType {
-public:
-  CustomType(size_t Val) : MVal(Val) {}
-  CustomType() : MVal(0) {}
-
-  bool operator<(const CustomType &RHS) const { return MVal < RHS.MVal; }
-  bool operator>(const CustomType &RHS) const { return MVal > RHS.MVal; }
-  bool operator==(const CustomType &RHS) const { return MVal == RHS.MVal; }
-
-private:
-  size_t MVal = 0;
-};
 
 #if VERSION == 1
 template <class CompT, class T> struct RadixSorterType;
@@ -87,28 +63,7 @@ template <> struct RadixSorterType<std::greater<CustomType>, CustomType> {
   using Type =
       oneapi_exp::radix_sorter<int, oneapi_exp::sorting_order::descending>;
 };
-#else
-template <class T> struct ConvertToSimpleType {
-  using Type = T;
-};
-
-// Dummy overloads for CustomType which is not supported by radix sorter
-template <> struct ConvertToSimpleType<CustomType> {
-  using Type = int;
-};
-
-template <class SorterT> struct ConvertToSortingOrder;
-
-template <class T> struct ConvertToSortingOrder<std::greater<T>> {
-  static const auto Type = oneapi_exp::sorting_order::descending;
-};
-
-template <class T> struct ConvertToSortingOrder<std::less<T>> {
-  static const auto Type = oneapi_exp::sorting_order::ascending;
-};
 #endif
-
-constexpr size_t ReqSubGroupSize = 8;
 
 template <UseGroupT UseGroup, int Dims, class T, class Compare>
 void RunJointSort(sycl::queue &Q, const std::vector<T> &DataToSort,
@@ -484,244 +439,11 @@ void RunSortOVerGroup(sycl::queue &Q, const std::vector<T> &DataToSort,
   }
 }
 
-#if VERSION == 2
-template <UseGroupT UseGroup, int Dims, class T, class U, class Compare>
-void RunSortKeyValueOverGroup(sycl::queue &Q, const std::vector<T> &DataToSort,
-                              const std::vector<U> &KeysToSort,
-                              const Compare &Comp) {
-
-  const size_t NumOfElements = DataToSort.size();
-  const size_t NumSubGroups = NumOfElements / ReqSubGroupSize + 1;
-
-  const sycl::nd_range<Dims> NDRange = [&]() {
-    if constexpr (Dims == 1)
-      return sycl::nd_range<1>{{NumOfElements}, {NumOfElements}};
-    else
-      return sycl::nd_range<2>{{1, NumOfElements}, {1, NumOfElements}};
-    static_assert(Dims < 3,
-                  "Only one and two dimensional kernels are supported");
-  }();
-
-  using RadixSorterT = oneapi_exp::radix_sorters::group_key_value_sorter<
-      typename ConvertToSimpleType<T>::Type, U,
-      ConvertToSortingOrder<Compare>::Type>;
-
-  std::size_t LocalMemorySizeDefault = 0;
-  std::size_t LocalMemorySizeRadix = 0;
-  if (UseGroup == UseGroupT::SubGroup) {
-    // Each sub-group needs a piece of memory for sorting
-    LocalMemorySizeDefault =
-        oneapi_exp::default_sorters::group_key_value_sorter<
-            T, U, Compare>::memory_required(sycl::memory_scope::sub_group,
-                                            ReqSubGroupSize);
-
-    LocalMemorySizeRadix = RadixSorterT::memory_required(
-        sycl::memory_scope::sub_group, ReqSubGroupSize);
-  } else {
-    // A single chunk of memory for each work-group
-    LocalMemorySizeDefault =
-        oneapi_exp::default_sorters::group_key_value_sorter<
-            T, U, Compare>::memory_required(sycl::memory_scope::work_group,
-                                            NumOfElements);
-
-    LocalMemorySizeRadix = RadixSorterT::memory_required(
-        sycl::memory_scope::work_group, NumOfElements);
-  }
-
-  std::vector<T> DataToSortCase0 = DataToSort;
-  std::vector<U> KeysToSortCase0 = KeysToSort;
-
-  std::vector<T> DataToSortCase1 = DataToSort;
-  std::vector<T> KeysToSortCase1 = KeysToSort;
-
-  std::vector<T> DataToSortCase2 = DataToSort;
-  std::vector<T> KeysToSortCase2 = KeysToSort;
-
-  std::vector<T> DataToSortCase3 = DataToSort;
-  std::vector<T> KeysToSortCase3 = KeysToSort;
-
-  // Sort data using 3 different versions of sort_over_group API
-  {
-    sycl::buffer<T> BufDataToSort0(DataToSortCase0.data(),
-                                   DataToSortCase0.size());
-    sycl::buffer<U> BufKeysToSort0(KeysToSortCase0.data(),
-                                   KeysToSortCase0.size());
-
-    sycl::buffer<T> BufDataToSort1(DataToSortCase1.data(),
-                                   DataToSortCase1.size());
-    sycl::buffer<T> BufKeysToSort1(KeysToSortCase1.data(),
-                                   KeysToSortCase1.size());
-
-    sycl::buffer<T> BufDataToSort2(DataToSortCase2.data(),
-                                   DataToSortCase2.size());
-    sycl::buffer<T> BufKeysToSort2(KeysToSortCase2.data(),
-                                   KeysToSortCase2.size());
-
-    sycl::buffer<T> BufDataToSort3(DataToSortCase3.data(),
-                                   DataToSortCase3.size());
-    sycl::buffer<T> BufKeysToSort3(KeysToSortCase3.data(),
-                                   KeysToSortCase3.size());
-
-    Q.submit([&](sycl::handler &CGH) {
-       auto AccDataToSort0 = sycl::accessor(BufDataToSort0, CGH);
-       auto AccKeysToSort0 = sycl::accessor(BufKeysToSort0, CGH);
-
-       auto AccDataToSort1 = sycl::accessor(BufDataToSort1, CGH);
-       auto AccKeysToSort1 = sycl::accessor(BufKeysToSort1, CGH);
-
-       auto AccDataToSort2 = sycl::accessor(BufDataToSort2, CGH);
-       auto AccKeysToSort2 = sycl::accessor(BufKeysToSort2, CGH);
-
-       auto AccDataToSort3 = sycl::accessor(BufDataToSort3, CGH);
-       auto AccKeysToSort3 = sycl::accessor(BufKeysToSort3, CGH);
-
-       // Allocate local memory for all sub-groups in a work-group
-       const size_t TotalLocalMemSizeDefault =
-           UseGroup == UseGroupT::SubGroup
-               ? LocalMemorySizeDefault * NumSubGroups
-               : LocalMemorySizeDefault;
-       sycl::local_accessor<std::byte, 1> ScratchDefault(
-           {TotalLocalMemSizeDefault}, CGH);
-
-       const size_t TotalLocalMemSizeRadix =
-           UseGroup == UseGroupT::SubGroup ? LocalMemorySizeRadix * NumSubGroups
-                                           : LocalMemorySizeRadix;
-
-       sycl::local_accessor<std::byte, 1> ScratchRadix({TotalLocalMemSizeRadix},
-                                                       CGH);
-
-       CGH.parallel_for<KernelNameKeyValueOverGroup<
-           IntWrapper<Dims>, UseGroupWrapper<UseGroup>, T, Compare>>(
-           NDRange, [=](sycl::nd_item<Dims> id) [[intel::reqd_sub_group_size(
-                        ReqSubGroupSize)]] {
-             const size_t GlobalLinearID = id.get_global_linear_id();
-
-             auto Group = [&]() {
-               if constexpr (UseGroup == UseGroupT::SubGroup)
-                 return id.get_sub_group();
-               else
-                 return id.get_group();
-             }();
-
-             // Each sub-group should use it's own part of the scratch pad
-             const size_t ScratchShiftDefault =
-                 UseGroup == UseGroupT::SubGroup
-                     ? id.get_sub_group().get_group_linear_id() *
-                           LocalMemorySizeDefault
-                     : 0;
-             std::byte *ScratchPtrDefault =
-                 &ScratchDefault[0] + ScratchShiftDefault;
-
-             if constexpr (std::is_same_v<Compare, std::less<T>>)
-               std::tie(AccKeysToSort0[GlobalLinearID],
-                        AccDataToSort0[GlobalLinearID]) =
-                   oneapi_exp::sort_key_value_over_group(
-                       oneapi_exp::group_with_scratchpad(
-                           Group, sycl::span{ScratchPtrDefault,
-                                             LocalMemorySizeDefault}),
-                       AccKeysToSort0[GlobalLinearID],
-                       AccDataToSort0[GlobalLinearID]); // (4)
-
-             std::tie(AccKeysToSort1[GlobalLinearID],
-                      AccDataToSort1[GlobalLinearID]) =
-                 oneapi_exp::sort_key_value_over_group(
-                     oneapi_exp::group_with_scratchpad(
-                         Group,
-                         sycl::span{ScratchPtrDefault, LocalMemorySizeDefault}),
-                     AccKeysToSort1[GlobalLinearID],
-                     AccDataToSort1[GlobalLinearID], Comp); // (5)
-
-             std::tie(AccKeysToSort2[GlobalLinearID],
-                      AccDataToSort2[GlobalLinearID]) =
-                 oneapi_exp::sort_key_value_over_group(
-                     Group, AccKeysToSort2[GlobalLinearID],
-                     AccDataToSort2[GlobalLinearID],
-                     oneapi_exp::default_sorters::group_key_value_sorter<
-                         T, U, Compare, /*ElementsPerWorkItem*/ 1>(sycl::span{
-                         ScratchPtrDefault, LocalMemorySizeDefault})); // (6)
-
-             // Each sub-group should use it's own part of the scratch pad
-             const size_t ScratchShiftRadix =
-                 UseGroup == UseGroupT::SubGroup
-                     ? id.get_sub_group().get_group_linear_id() *
-                           LocalMemorySizeRadix
-                     : 0;
-             std::byte *ScratchPtrRadix = &ScratchRadix[0] + ScratchShiftRadix;
-
-             // Radix doesn't support custom types
-             if constexpr (!std::is_same_v<CustomType, T>)
-               std::tie(AccKeysToSort3[GlobalLinearID],
-                        AccDataToSort3[GlobalLinearID]) =
-                   oneapi_exp::sort_key_value_over_group(
-                       Group, AccKeysToSort3[GlobalLinearID],
-                       AccDataToSort3[GlobalLinearID],
-                       RadixSorterT(
-                           sycl::span{ScratchPtrRadix,
-                                      LocalMemorySizeRadix})); // (6) radix
-           });
-     }).wait_and_throw();
-  }
-
-  // Verification
-  {
-    std::vector<std::pair<T, T>> KeyDataToSort;
-    KeyDataToSort.reserve(KeysToSort.size());
-    std::transform(KeysToSort.begin(), KeysToSort.end(), DataToSort.begin(),
-                   std::back_inserter(KeyDataToSort),
-                   [](T Key, T Value) { return std::make_pair(Key, Value); });
-    // Emulate independent sorting of each work-group/sub-group
-    const size_t ChunkSize = UseGroup == UseGroupT::SubGroup
-                                 ? ReqSubGroupSize
-                                 : NDRange.get_local_range().size();
-    auto It = KeyDataToSort.begin();
-    auto KeyValueComp = [&](const std::pair<T, T> &A,
-                            const std::pair<T, T> &B) -> bool {
-      return Comp(A.first, B.first);
-    };
-    for (; (It + ChunkSize) < KeyDataToSort.end(); It += ChunkSize)
-      std::stable_sort(It, It + ChunkSize, KeyValueComp);
-
-    // Sort remainder
-    std::stable_sort(It, KeyDataToSort.end(), KeyValueComp);
-
-    std::vector<T> KeysSorted, DataSorted;
-    KeysSorted.reserve(KeyDataToSort.size());
-    DataSorted.reserve(KeyDataToSort.size());
-    std::transform(
-        KeyDataToSort.begin(), KeyDataToSort.end(),
-        std::back_inserter(KeysSorted),
-        [](const std::pair<T, T> &KeyValue) { return KeyValue.first; });
-    std::transform(
-        KeyDataToSort.begin(), KeyDataToSort.end(),
-        std::back_inserter(DataSorted),
-        [](const std::pair<T, T> &KeyValue) { return KeyValue.second; });
-
-    if constexpr (std::is_same_v<Compare, std::less<T>>) {
-      assert(KeysToSortCase0 == KeysSorted);
-      assert(DataToSortCase0 == DataSorted);
-    }
-
-    assert(KeysToSortCase1 == KeysSorted);
-    assert(DataToSortCase1 == DataSorted);
-    assert(KeysToSortCase2 == KeysSorted);
-    assert(DataToSortCase2 == DataSorted);
-    if constexpr (!std::is_same_v<CustomType, T>) {
-      assert(KeysToSortCase3 == KeysSorted);
-      assert(DataToSortCase3 == DataSorted);
-    }
-  }
-}
-#endif
-
 template <class T> void RunOverType(sycl::queue &Q, size_t DataSize) {
   std::vector<T> DataReversed(DataSize);
-  std::vector<T> KeysReversed(DataSize);
-
   std::vector<T> DataRandom(DataSize);
-  std::vector<T> KeysRandom(DataSize);
 
   std::iota(DataReversed.rbegin(), DataReversed.rend(), (size_t)0);
-  KeysReversed = DataReversed;
 
   // Fill using random numbers
   {
@@ -729,23 +451,12 @@ template <class T> void RunOverType(sycl::queue &Q, size_t DataSize) {
     std::normal_distribution<float> distribution((10.0), (2.0));
     for (T &Elem : DataRandom)
       Elem = T(distribution(generator));
-
-    for (T &Elem : KeysRandom)
-      Elem = T(distribution(generator));
   }
 
   auto RunOnDataAndComp = [&](const std::vector<T> &Data,
-                              const std::vector<T> &Keys,
                               const auto &Comparator) {
     RunSortOVerGroup<UseGroupT::WorkGroup, 1>(Q, Data, Comparator);
     RunSortOVerGroup<UseGroupT::WorkGroup, 2>(Q, Data, Comparator);
-
-#if VERSION == 2
-    RunSortKeyValueOverGroup<UseGroupT::WorkGroup, 1>(Q, Data, Keys,
-                                                      Comparator);
-    RunSortKeyValueOverGroup<UseGroupT::WorkGroup, 2>(Q, Data, Keys,
-                                                      Comparator);
-#endif
 
     RunJointSort<UseGroupT::WorkGroup, 1>(Q, Data, Comparator);
     RunJointSort<UseGroupT::WorkGroup, 2>(Q, Data, Comparator);
@@ -759,19 +470,14 @@ template <class T> void RunOverType(sycl::queue &Q, size_t DataSize) {
     RunSortOVerGroup<UseGroupT::SubGroup, 1>(Q, Data, Comparator);
     RunSortOVerGroup<UseGroupT::SubGroup, 2>(Q, Data, Comparator);
 
-#if VERSION == 2
-    RunSortKeyValueOverGroup<UseGroupT::SubGroup, 1>(Q, Data, Keys, Comparator);
-    RunSortKeyValueOverGroup<UseGroupT::SubGroup, 2>(Q, Data, Keys, Comparator);
-#endif
-
     RunJointSort<UseGroupT::SubGroup, 1>(Q, Data, Comparator);
     RunJointSort<UseGroupT::SubGroup, 2>(Q, Data, Comparator);
   };
 
-  RunOnDataAndComp(DataReversed, KeysReversed, std::greater<T>{});
-  RunOnDataAndComp(DataReversed, KeysReversed, std::less<T>{});
-  RunOnDataAndComp(DataRandom, KeysRandom, std::less<T>{});
-  RunOnDataAndComp(DataRandom, KeysRandom, std::greater<T>{});
+  RunOnDataAndComp(DataReversed, std::greater<T>{});
+  RunOnDataAndComp(DataReversed, std::less<T>{});
+  RunOnDataAndComp(DataRandom, std::less<T>{});
+  RunOnDataAndComp(DataRandom, std::greater<T>{});
 }
 
 int main() {
